+++ util/src/main/scala/org/opentorah/util/Strings.scala
@@ -1,5 +1,8 @@
 package org.opentorah.util
 
+import java.io.{ByteArrayInputStream, InputStream}
+import java.nio.charset.Charset
+
 object Strings {
 
   def split(what: String, on: Char): (String, Option[String]) = what.lastIndexOf(on) match {
@@ -39,4 +42,12 @@ object Strings {
 
   def empty2none(string: String): Option[String] =
     if (string == null || string.isEmpty) None else Some(string)
+
+  def utf8: Charset = Charset.forName("UTF-8")
+
+  def string2stream(string: String): InputStream = new ByteArrayInputStream(string.getBytes(utf8))
+
+  def drop(from: String, prefix: String): String =
+    if (from.startsWith(prefix)) from.substring(prefix.length)
+    else throw new IllegalArgumentException(s"String '$from' doesn't start with '$prefix'")
 }